+++ admin/modulos/sistemas-de-pagos/sistemas_de_pagos.php
@@ -77,16 +77,19 @@ include('../../includes/header.php');
                     (<?= htmlspecialchars($m['tipo']) ?>):
                     <?= htmlspecialchars($m['valor']) ?>
 
-                    <a href="editar_pago.php?id=<?= $m['id'] ?>" class="btn btn-ver">
-                        <i class="fa-solid fa-pen"></i>
-                        Editar
-                    </a>
+                    <?php if (esAdmin()): ?>
+                        <a href="editar_pago.php?id=<?= $m['id'] ?>" class="btn btn-ver">
+                            <i class="fa-solid fa-pen"></i>
+                            Editar
+                        </a>
 
-                    <a href="eliminar_pago.php?id=<?= $m['id'] ?>" 
-                        class="btn btn-borrar" 
-                        onclick="return confirm('¿Seguro que deseas eliminar este método de pago?');">
-                        <i class="fa-solid fa-trash"></i> Borrar
-                    </a>
+                        <a href="eliminar_pago.php?id=<?= $m['id'] ?>"
+                            class="btn btn-borrar"
+                            onclick="return confirm('¿Seguro que deseas eliminar este método de pago?');">
+                            <i class="fa-solid fa-trash"></i> Borrar
+                        </a>
+                    <?php endif; ?>
+
                 </li>
             <?php endforeach; ?>
         </ul>
